+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/GravityDefying.java
@@ -1,21 +1,15 @@
 package insane96mcp.iguanatweaksexpanded.module.experience.enchantments.enchantment;
 
-import insane96mcp.iguanatweaksexpanded.module.experience.enchantments.NewEnchantmentsFeature;
+import insane96mcp.iguanatweaksexpanded.IguanaTweaksExpanded;
 import net.minecraft.world.entity.EquipmentSlot;
-import net.minecraft.world.entity.ai.attributes.AttributeModifier;
-import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
-import net.minecraftforge.common.ForgeMod;
-import net.minecraftforge.event.ItemAttributeModifierEvent;
-import net.minecraftforge.event.entity.living.LivingFallEvent;
 
-import java.util.UUID;
+public class GravityDefying extends Enchantment {
+    public static final String NBT_TAG = IguanaTweaksExpanded.RESOURCE_PREFIX + "gravity_defying";
 
-public class GravityDefying extends Enchantment implements IAttributeEnchantment {
-    public static final UUID GRAVITY_MODIFIER_UUID = UUID.fromString("7567b3ad-a4c6-4700-8bf0-cd8c4356c155");
     public GravityDefying() {
-        super(Rarity.RARE, EnchantmentCategory.ARMOR_FEET, new EquipmentSlot[] {EquipmentSlot.FEET});
+        super(Rarity.VERY_RARE, EnchantmentCategory.BOW, new EquipmentSlot[] {EquipmentSlot.MAINHAND, EquipmentSlot.OFFHAND});
     }
 
     @Override
@@ -26,35 +20,5 @@ public class GravityDefying extends Enchantment implements IAttributeEnchantment
     @Override
     public int getMaxCost(int level) {
         return this.getMinCost(level) + 20;
-    }
-
-    public static float getGravityReduction() {
-        return 0.28f;
-    }
-
-    @Override
-    public boolean canApplyAtEnchantingTable(ItemStack stack) {
-        return false;
-    }
-
-    @Override
-    public boolean canEnchant(ItemStack pStack) {
-        return false;
-    }
-
-    @Override
-    public boolean isDiscoverable() {
-        return false;
-    }
-
-    public static void applyFallDamageReduction(LivingFallEvent event) {
-        if (event.getEntity().getItemBySlot(EquipmentSlot.FEET).getEnchantmentLevel(NewEnchantmentsFeature.GRAVITY_DEFYING.get()) <= 0)
-            return;
-        event.setDistance((event.getDistance() - 0.5f));
-    }
-
-    @Override
-    public void applyAttributeModifier(ItemAttributeModifierEvent event, int enchantmentLvl) {
-        event.addModifier(ForgeMod.ENTITY_GRAVITY.get(), new AttributeModifier(GRAVITY_MODIFIER_UUID, "Gravity Defying enchantment", -getGravityReduction(), AttributeModifier.Operation.MULTIPLY_BASE));
     }
 }